+++ Network Mom ACL Analyzer/enumerations/IosToken.swift
@@ -20,7 +20,7 @@ enum IosToken: Equatable {
     case comment
     case log
     case established
-    case fourOctet(UInt)
+    case fourOctet(UInt128)
     case number(UInt)
     case name(String)
