+++ Network Mom ACL Analyzer/enumerations/IosXrToken.swift
@@ -21,7 +21,7 @@ enum IosXrToken: Equatable {
     case comment
     case log
     case established
-    case fourOctet(UInt)
+    case fourOctet(UInt128)
     case cidr(IpRange)
     case number(UInt)
     case name(String)
