+++ Network Mom ACL Analyzer/extensions/String+extensions.swift
@@ -9,14 +9,14 @@
 import Foundation
 
 extension String {
-    var ipv4address: UInt? {
+    var ipv4address: UInt128? {
         let octets = self.split(separator: ".")
         guard octets.count == 4 else { return nil }
         guard let octet1 = UInt8(octets[0]) else { return nil }
         guard let octet2 = UInt8(octets[1]) else { return nil }
         guard let octet3 = UInt8(octets[2]) else { return nil }
         guard let octet4 = UInt8(octets[3]) else { return nil }
-        let answer: UInt = UInt(octet1) * 256 * 256 * 256 + UInt(octet2) * 256 * 256 + UInt(octet3) * 256 + UInt(octet4)
+        let answer: UInt128 = UInt128(octet1) * 256 * 256 * 256 + UInt128(octet2) * 256 * 256 + UInt128(octet3) * 256 + UInt128(octet4)
         return answer
     }
